+++ src/services/inductee-sheets.service.ts
@@ -1,11 +1,11 @@
 import { z } from "zod";
 
+import { RowWiseSheetsService } from "../abc/sheets.abc";
 import { GoogleSheetsClient } from "../clients/sheets.client";
 import env from "../env";
 import { cleanProvidedUsername } from "../features/inductee-role/input.utils";
 import type { Seconds } from "../types/branded.types";
 import { SystemDateClient } from "../utils/date.utils";
-import { RowWiseSheetsService } from "./sheets.service";
 
 export enum UpeMajor {
   Cs = "Computer Science",
